fix moved package of TestLoggerExtension
